backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
pydantic-settings
openai-whisper # Using OpenAI Whisper for ASR
langchain
langchain-community
fpdf2
soundfile # Needed for audio loading
# alibabacloud_nls # Removed
# nemo_toolkit[asr] # Removed, using Whisper instead
# torch # Keep separate install for now (Whisper might need it)
# torchaudio # Keep separate install for now (Whisper might need it)
